+++ doc/conf.py
@@ -33,7 +33,7 @@ on_rtd = os.environ.get("READTHEDOCS", None) == "True"
 
 import sys
 sys.path.append('.')
-import create_graph
+#import create_graph
 
 
 import pkg_resources